Test added refactor some part

+++ src/main/java/nl/group9/xmlanalyser/service/serviceimpl/RabbitMQServiceImpl.java
@@ -1,6 +1,5 @@
 package nl.group9.xmlanalyser.service.serviceimpl;
 
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import nl.group9.xmlanalyser.entity.Xml;
 import nl.group9.xmlanalyser.service.RabbitMQService;
@@ -12,11 +11,13 @@ import org.apache.commons.io.FileUtils;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
+
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Date;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +28,8 @@ public class RabbitMQServiceImpl implements RabbitMQService {
 
     @Override
     public XmlResponse processXml(XmlRequest xmlRequest) {
+
+        Optional.ofNullable(xmlRequest.getUrl()).orElseThrow(() -> new RuntimeException("Url can not be null"));
 
         Xml xml = new Xml();
         xml.setAnalyseDate(new Date());
@@ -47,15 +50,19 @@ public class RabbitMQServiceImpl implements RabbitMQService {
     @Override
     public void receiveXml(Xml xml) {
 
+        File file = null;
         try {
-            File file = new File(xml.getId() + ".xml");
+            file = new File(xml.getId() + ".xml");
             FileUtils.copyURLToFile(new URL(xml.getUrl()), file);
             xmlService.parseXml(xml, file);
-            file.delete();
         } catch (MalformedURLException me) {
             me.printStackTrace();
         } catch (IOException ioe) {
             ioe.printStackTrace();
+        } finally {
+            if (file != null) {
+                file.delete();
+            }
         }
     }
 }
